requirements.txt
python-dotenv
langchain
PyPDF2
faiss-cpu
streamlit
openai
langchain-openai
-e .